// src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      listing_access_control: {
        Row: {
          actioned_at: string | null
          actioned_by_user: string | null
          created_at: string
          id: string
          listing: string
          request_status: Database["public"]["Enums"]["request_status"]
          requested_by_firm: string
          requested_by_user: string
        }
        Insert: {
          actioned_at?: string | null
          actioned_by_user?: string | null
          created_at?: string
          id?: string
          listing: string
          request_status?: Database["public"]["Enums"]["request_status"]
          requested_by_firm: string
          requested_by_user: string
        }
        Update: {
          actioned_at?: string | null
          actioned_by_user?: string | null
          created_at?: string
          id?: string
          listing?: string
          request_status?: Database["public"]["Enums"]["request_status"]
          requested_by_firm?: string
          requested_by_user?: string
        }
        Relationships: [
          {
            foreignKeyName: "listing_access_control_listing_id_fkey"
            columns: ["listing"]
            isOneToOne: false
            referencedRelation: "listing"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "listing_access_control_requested_by_firm_id_fkey"
            columns: ["requested_by_firm"]
            isOneToOne: false
            referencedRelation: "firm"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "listing_access_control_requested_by_user_id_fkey"
            columns: ["requested_by_user"]
            isOneToOne: false
            referencedRelation: "user"
            referencedColumns: ["id"]
          },
        ]
      }
      listing_content: {
        Row: {
          content_type: Database["public"]["Enums"]["listing_type"]
          created_at: string
          description: string
          entity_type: Database["public"]["Enums"]["entity_type"][]
          id: string
          name: string
          owned_by_firm: string
          region: Database["public"]["Enums"]["region"]
          workpaper_type: Database["public"]["Enums"]["workpaper_type"][]
        }
        Insert: {
          content_type: Database["public"]["Enums"]["listing_type"]
          created_at?: string
          description: string
          entity_type: Database["public"]["Enums"]["entity_type"][]
          id?: string
          name: string
          owned_by_firm?: string
          region: Database["public"]["Enums"]["region"]
          workpaper_type: Database["public"]["Enums"]["workpaper_type"][]
        }
        Update: {
          content_type?: Database["public"]["Enums"]["listing_type"]
          created_at?: string
          description?: string
          entity_type?: Database["public"]["Enums"]["entity_type"][]
          id?: string
          name?: string
          owned_by_firm?: string
          region?: Database["public"]["Enums"]["region"]
          workpaper_type?: Database["public"]["Enums"]["workpaper_type"][]
        }
        Relationships: [
          {
            foreignKeyName: "listing_content_owned_by_firm_fkey"
            columns: ["owned_by_firm"]
            isOneToOne: false
            referencedRelation: "firm"
            referencedColumns: ["id"]
          },
        ]
      }
      saved_listing: {
        Row: {
          created_at: string
          id: string
          listing: string
          saved_by_firm: string
          saved_by_user: string
        }
        Insert: {
          created_at?: string
          id?: string
          listing: string
          saved_by_firm: string
          saved_by_user: string
        }
        Update: {
          created_at?: string
          id?: string
          listing?: string
          saved_by_firm?: string
          saved_by_user?: string
        }
        Relationships: [
          {
            foreignKeyName: "saved_listing_listing_id_fkey"
            columns: ["listing"]
            isOneToOne: false
            referencedRelation: "listing"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "saved_listing_saved_by_firm_fkey"
            columns: ["saved_by_firm"]
            isOneToOne: false
            referencedRelation: "firm"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "saved_listing_saved_by_user_fkey"
            columns: ["saved_by_user"]
            isOneToOne: false
            referencedRelation: "user"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      custom_access_token_hook: { Args: { event: Json }; Returns: Json }
    }
    Enums: {
      entity_type: "company" | "individual" | "partnership" | "trust"
      listing_status: "active" | "deleted"
      listing_type:
        | "calculation"
        | "checklist"
        | "report"
        | "procedure"
        | "wiki"
      listing_visibility: "public" | "request_access"
      region: "australia" | "newZealand" | "unitedKingdom" | "republicOfIreland"
      request_status: "pending" | "approved" | "rejected"
      vendor_status: "active" | "blocked"
      workpaper_type: "compliance" | "itr" | "bas" | "taxPlanning" | "fbt"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never